adjusted useefect from Navbar and added server Url to http clirnt and confug
