+++ client-side/src/Services/config.js
@@ -1,6 +1,8 @@
 import axios from "axios";
+const server_URL = import.meta.env.VITE_SERVER_URL;
+
 const httpAuth = axios.create({
-  baseURL: "http://localhost:5000/api/v1/",
+  baseURL: `${server_URL}/api/v1/`,
   headers: {
     "Content-Type": "application/json",
   },
